chore(layerstyle): code cleanup and review
Performed basic code cleanup on LayerStyle.h
Updated unit tests accordingly.
Fixing LayerStyle case-insensitive issue.
