+++ lib/LayerStyle.cpp
@@ -4,6 +4,51 @@
 #include "LayerStyle.h"
 
 /*!
+ * \brief getColorFromString creates a QColor object from an HSL color string.
+ *
+ * The string is expected to be in one of the following formats:
+ *      "hsl(hue, stauration%, lightness%)"
+ *      "hsla(hue, stauration%, lightness%, alpha)"
+ *
+ * \param colorString is a QString containing color data
+ *
+ * \return a QColor object.
+ */
+QColor Bach::getColorFromString(QString colorString)
+{
+    colorString.remove(" ");
+    //All parameters for QColor::fromHslF need to be between 0 and 1.
+    if (colorString.startsWith("hsl(")) {
+        static QRegularExpression re { ".*\\((\\d+),(\\d+)%,(\\d+)%\\)" };
+        QRegularExpressionMatch match = re.match(colorString);
+        if (match.capturedTexts().length() >= 4) {
+            return QColor::fromHslF(match.capturedTexts().at(1).toInt()/359.,
+                                    match.capturedTexts().at(2).toInt()/100.,
+                                    match.capturedTexts().at(3).toInt()/100.);
+        }
+    }
+    if (colorString.startsWith("hsla(")) {
+        static QRegularExpression re { ".*\\((\\d+),(\\d+)%,(\\d+)%,(\\d?\\.?\\d*)\\)" };
+        QRegularExpressionMatch match = re.match(colorString);
+        if (match.capturedTexts().length() >= 5) {
+            return QColor::fromHslF(match.capturedTexts().at(1).toInt()/359.,
+                                    match.capturedTexts().at(2).toInt()/100.,
+                                    match.capturedTexts().at(3).toInt()/100.,
+                                    match.capturedTexts().at(4).toFloat());
+        }
+    }
+
+    // In case the color has a different format than expected.
+    // If the format cannot be handeled by QColor, we return black as a default color.
+    QColor returnColor = QColor::fromString(colorString);
+    if (!returnColor.isValid()) {
+        return Qt::black;
+    } else {
+        return returnColor;
+    }
+}
+
+/*!
  * \brief AbstractLayerStyle::fromJson parses different layer style types.
  *
  * The layerType data type is converted to strings.
@@ -22,64 +67,38 @@
 std::unique_ptr<AbstractLayerStyle> AbstractLayerStyle::fromJson(const QJsonObject &json)
 {
     QString layerType = json.value("type").toString();
+
     std::unique_ptr<AbstractLayerStyle> returnLayerPtr;
+
     if (layerType == "background") {
         returnLayerPtr = BackgroundStyle::fromJson(json);
-    }else if ( layerType == "fill") {
+    } else if (layerType == "fill") {
         returnLayerPtr = FillLayerStyle::fromJson(json);
-    }else if (layerType == "line") {
+    } else if (layerType == "line") {
         returnLayerPtr =  LineLayerStyle::fromJson(json);
-    }else if (layerType == "symbol") {
+    } else if (layerType == "symbol") {
         returnLayerPtr = SymbolLayerStyle::fromJson(json);
     } else {
         returnLayerPtr = NotImplementedStyle::fromJson(json);
     }
-    AbstractLayerStyle *newLayer = returnLayerPtr.get();
-    newLayer->m_id = json.value("id").toString();
-    newLayer->m_source = json.value("source").toString();
-    newLayer->m_sourceLayer = json.value("source-layer").toString();
-    newLayer->m_minZoom = json.value("minzoom").toInt(0);
-    newLayer->m_maxZoom = json.value("maxzoom").toInt(24);
+
+    AbstractLayerStyle &newLayer = *returnLayerPtr.get();
+    newLayer.m_id = json.value("id").toString();
+    newLayer.m_source = json.value("source").toString();
+    newLayer.m_sourceLayer = json.value("source-layer").toString();
+    newLayer.m_minZoom = json.value("minzoom").toInt(0);
+    newLayer.m_maxZoom = json.value("maxzoom").toInt(24);
     QJsonValue layout = json.value("layout");
     if (layout != QJsonValue::Undefined) {
-        newLayer->m_visibility = (layout.toObject().contains("visibility")) ? layout.toObject().value("visibility").toString() : "none";
+        newLayer.m_visibility = (layout.toObject().contains("visibility")) ? layout.toObject().value("visibility").toString() : "none";
     } else {
-        newLayer->m_visibility = QString("none");
+        newLayer.m_visibility = QString("none");
     }
 
     if (json.contains("filter")) {
-        newLayer->m_filter = json.value("filter").toArray();
+        newLayer.m_filter = json.value("filter").toArray();
     }
     return returnLayerPtr;
-}
-
-/* Stylesheet parsing functions are documented and implemented below. */
-
-/*!
- * \brief StyleSheet::~StyleSheet is the StyleSheet class destructor.
- */
-StyleSheet::~StyleSheet()
-{
-}
-
-/*!
- * \brief StyleSheet::parseSheet parses a style sheet and populates the
- * layers array of a styleSheetObject.
- *
- * \param styleSheet is a style sheet to parse, passed as a reference to
- * a QJsonDocument.
- */
-void StyleSheet::parseSheet(const QJsonDocument &styleSheet)
-{
-   QJsonObject styleSheetObject = styleSheet.object();
-    m_id = styleSheetObject.value("id").toString();
-    m_version = styleSheetObject.value("version").toInt();
-    m_name = styleSheetObject.value("name").toString();
-
-    QJsonArray layers = styleSheetObject.value("layers").toArray();
-    for (const auto &layer : layers) {
-        m_layerStyles.push_back(AbstractLayerStyle::fromJson(layer.toObject()));
-    }
 }
 
 /*!
@@ -90,9 +109,46 @@ void StyleSheet::parseSheet(const QJsonDocument &styleSheet)
  *
  * \return  is either the parsed data or a nullopt of there was no data.
  */
-std::optional<StyleSheet> StyleSheet::fromJson(const QJsonDocument &input)
+std::optional<StyleSheet> StyleSheet::fromJson(const QJsonDocument &styleSheetJson)
 {
     StyleSheet out;
-    out.parseSheet(input);
+
+    if (!styleSheetJson.isObject()) {
+        return std::nullopt;
+    }
+    QJsonObject styleSheetObject = styleSheetJson.object();
+
+    {
+        // Load id of the stylesheet.
+        auto idIter = styleSheetObject.find("id");
+        if (idIter == styleSheetObject.end()) {
+            return std::nullopt;
+        }
+        out.m_id = idIter->toString();
+    }
+
+    {
+        // Load version of stylesheet.
+        auto versionIter = styleSheetObject.find("version");
+        if (versionIter == styleSheetObject.end() || !versionIter->isDouble()) {
+            return std::nullopt;
+        }
+        out.m_version = versionIter->toInt();
+    }
+
+    {
+        // Load version of stylesheet.
+        auto nameIter = styleSheetObject.find("name");
+        if (nameIter == styleSheetObject.end() || !nameIter->isString()) {
+            return std::nullopt;
+        }
+        out.m_name = nameIter->toString();
+    }
+
+    QJsonArray layers = styleSheetObject.value("layers").toArray();
+    for (QJsonValueConstRef layer : layers) {
+        out.m_layerStyles.push_back(AbstractLayerStyle::fromJson(layer.toObject()));
+    }
+
     return out;
 }
